refactor: move CLI workflow input expansion to runtime
