+++ packages/cli/src/commands/multiAgent.ts
@@ -69,7 +69,7 @@ import {
 import {
   createStdinWorkflowInputMaterializer,
   expandRunInputs,
-} from './_helpers/workflowInputs';
+} from '../runtime/workflowInputs';
 
 type CliToolUseRunResult = Extract<CliRunResult, { category: 'toolUse' }>;
