+++ src/test-kernel/cli/MultiAgentRunCommand.vitest.mts
@@ -122,7 +122,7 @@ vi.mock('@cli/commands/_helpers/runExecution', () => ({
   executeCliRequest: mocks.executeCliRequest,
 }));
 
-vi.mock('@cli/commands/_helpers/workflowInputs', () => ({
+vi.mock('@cli/runtime/workflowInputs', () => ({
   createStdinWorkflowInputMaterializer: vi.fn(() => mocks.stdinInputFile),
   expandRunInputs: mocks.expandRunInputs,
 }));
